Filtrado de tipos para insercion y actualizacion

+++ src/modules/controller/users/user.gateway.js
@@ -12,6 +12,8 @@ const findAll = async () => {
 }
 
 const findById = async (id) => {
+    if (!id) throw Error('Missing fields');
+    if (Number.isNaN(id)) throw Error('Wrong type');
     try {
         const sql = `SELECT * FROM users WHERE id=?`;
         const user = await query(sql, [id]);
@@ -23,6 +25,9 @@ const findById = async (id) => {
 }
 
 const save = async (user) => {
+    if (!user) throw Error('Missing fields');
+    if (user.role != 'admin' && user.role != 'user') throw Error('Wrong type of role');
+    if (user.status != 1 && user.status != 0) throw Error('Wrong type status');
     try {
         const { email, password, role, personal_id } = user;
         // make an insert query with the status in 1
@@ -36,6 +41,11 @@ const save = async (user) => {
 }
 
 const update = async (user, id) => {
+    if (!user) throw Error('Missing fields (user)');
+    if (!id) throw Error('Missing fields (id)');
+    if (Number.isNaN(id)) throw Error('Wrong type');
+    if (user.role != 'admin' && user.role != 'user') throw Error('Wrong type of role');
+    if (user.status != 1 && user.status != 0) throw Error('Wrong type status');
     try {
         const { email, password, role, status, personal_id } = user;
         const sql = `UPDATE users SET email=?, password=?, role=?, status=?, personal_id=? WHERE id=?`;
@@ -48,6 +58,8 @@ const update = async (user, id) => {
 }
 
 const remove = async (id) => {
+    if (!id) throw Error('Missing fields');
+    if (Number.isNaN(id)) throw ("Wrong Type");
     try {
         const sql = `DELETE FROM users WHERE id=?`;
         await query(sql, [id]);
